增强双流模型，增加验证指标与日志记录功能。添加了验证轮次函数以计算各项指标，更新了训练循环以记录每个轮次的详细信息，并优化了评估过程中的边界框处理。同时，提升了数据集加载器中数据类型的一致性，并针对有向目标检测（OBB）调整了模型初始化方式。

+++ ultralytics/data/dataset_obb.py
@@ -499,10 +499,10 @@ class DualStreamOBBDataLoader:
         
         if labels:
             # Concatenate all labels
-            labels = torch.from_numpy(np.concatenate(labels, axis=0))
+            labels = torch.from_numpy(np.concatenate(labels, axis=0)).float()
         else:
             # Empty batch: [batch_idx, cls, 5 coordinates (xywhr)] = 7 columns
-            labels = torch.zeros(0, 7)
+            labels = torch.zeros(0, 7, dtype=torch.float32)
         
         return {
             "vis": vis_imgs,
